# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fix_Smart_CMS API",
    "version": "1.0.3",
    "description": "\n        Comprehensive API documentation for Fix_Smart_CMS (NLC-CMS) - A modern complaint management system.\n        \n        ## Features\n        - **Multi-role Authentication**: Citizens, Ward Officers, Maintenance Teams, Administrators\n        - **Complaint Management**: Complete lifecycle from registration to resolution\n        - **Geographic Routing**: Ward-based automatic complaint assignment\n        - **File Management**: Unified attachment system for all file types\n        - **Real-time Notifications**: Email and in-app notification system\n        - **Analytics & Reporting**: Comprehensive dashboard and reporting capabilities\n        \n        ## Authentication\n        Most endpoints require JWT authentication. Include the token in the Authorization header:\n        `Authorization: Bearer <your-jwt-token>`\n        \n        ## Rate Limiting\n        API requests are rate-limited to prevent abuse. Development environment has higher limits.\n        \n        ## Response Format\n        All API responses follow a consistent format:\n        ```json\n        {\n          \"success\": boolean,\n          \"message\": \"string\",\n          \"data\": object | array | null,\n          \"errors\": array (optional)\n        }\n        ```\n      ",
    "contact": {
      "name": "API Support",
      "email": "api-support@nlc-cms.gov.in",
...
    },
...
    },
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "ComplaintType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": 1,
            "description": "Unique identifier for the complaint type"
          },
          "name": {
            "type": "string",
            "example": "STREET_LIGHTING",
            "description": "Name of the complaint type"
          },
          "description": {
            "type": "string",
            "example": "Issues related to street lighting",
            "description": "Description of the complaint type"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "example": "MEDIUM",
            "description": "Default priority level"
          },
          "slaHours": {
            "type": "integer",
            "description": "Service Level Agreement in hours",
            "example": 48
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Whether the complaint type is active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
...
      "GuestComplaintRequest": {
        "type": "object",
        "required": [
          "description",
          "area",
          "contactPhone",
          "wardId"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Brief title of the complaint",
            "example": "Street light not working"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the complaint",
            "example": "The street light on MG Road has been non-functional for 3 days"
          },
          "type": {
            "type": "string",
            "description": "Type of complaint",
            "example": "STREET_LIGHTING"
          },
          "area": {
            "type": "string",
            "description": "Area where the complaint is located",
            "example": "MG Road"
          },
          "landmark": {
            "type": "string",
            "description": "Nearby landmark",
            "example": "Near City Mall"
          },
          "address": {
            "type": "string",
            "description": "Full address",
            "example": "MG Road, Near City Mall, Kochi"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude coordinate",
            "example": 9.9312
          },
          "longitude": {
            "type": "number",
            "description": "Longitude coordinate",
            "example": 76.2673
          },
          "contactName": {
            "type": "string",
            "description": "Contact person name",
            "example": "John Doe"
          },
          "contactEmail": {
            "type": "string",
            "format": "email",
            "description": "Contact email address",
            "example": "john.doe@example.com"
          },
          "contactPhone": {
            "type": "string",
            "description": "Contact phone number",
            "example": "+91-9876543210"
          },
          "wardId": {
            "type": "string",
            "description": "Ward ID where complaint is located",
            "example": "ward123"
          },
          "subZoneId": {
            "type": "string",
            "description": "Sub-zone ID (optional)",
            "example": "subzone123"
          },
          "isAnonymous": {
            "type": "boolean",
            "description": "Whether to submit anonymously",
            "default": false
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "default": "MEDIUM"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            },
            "description": "File attachments (max 5 files, 10MB each)"
          }
        }
      },
...
      "OTPVerificationRequest": {
        "type": "object",
        "required": [
          "email",
          "otpCode"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "otpCode": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "example": "123456"
          }
        }
      },
...
    },
...
  },
...
  "paths": {
...
    "/api/complaint-types": {
      "get": {
        "summary": "Get all complaint types",
        "tags": [
          "Complaint Types"
        ],
        "responses": {
          "200": {
            "description": "List of all complaint types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ComplaintType"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new complaint type",
        "tags": [
          "Complaint Types"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintType"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Complaint type created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/complaint-types/stats": {
      "get": {
        "summary": "Get complaint type statistics",
        "tags": [
          "Complaint Types"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Complaint type usage statistics"
          }
        }
      }
    },
...
    "/api/guest/complaint": {
      "post": {
        "summary": "Submit a guest complaint",
        "tags": [
          "Guest"
        ],
        "description": "Submit a complaint as a guest user without authentication",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/GuestComplaintRequest"
              },
              "encoding": {
                "attachments": {
                  "contentType": "image/*, application/pdf, application/msword, text/plain"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Complaint submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestComplaintResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "413": {
            "description": "File too large (max 10MB per file)"
          },
          "415": {
            "description": "Unsupported file type"
          }
        }
      }
    },
    "/api/guest/complaint-with-attachments": {
      "post": {
        "summary": "Submit guest complaint with multiple attachments",
        "tags": [
          "Guest"
        ],
        "description": "Submit a complaint with multiple file attachments",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/GuestComplaintRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Complaint with attachments submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestComplaintResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "413": {
            "description": "File too large"
          },
          "415": {
            "description": "Unsupported file type"
          }
        }
      }
    },
    "/api/guest/verify-otp": {
      "post": {
        "summary": "Verify OTP and register guest complaint",
        "tags": [
          "Guest"
        ],
        "description": "Verify OTP sent to email and complete complaint registration",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/OTPVerificationRequest"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "attachments": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "binary"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified and complaint registered successfully"
          },
          "400": {
            "description": "Invalid or expired OTP"
          },
          "404": {
            "description": "OTP session not found"
          }
        }
      }
    },
    "/api/guest/resend-otp": {
      "post": {
        "summary": "Resend OTP for guest complaint verification",
        "tags": [
          "Guest"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP resent successfully"
          },
          "400": {
            "description": "Invalid email or no pending OTP session"
          },
          "429": {
            "description": "Too many OTP requests"
          }
        }
      }
    },
    "/api/guest/track/{complaintId}": {
      "get": {
        "summary": "Track complaint status",
        "tags": [
          "Guest"
        ],
        "description": "Track the status of a complaint using complaint ID",
        "parameters": [
          {
            "in": "path",
            "name": "complaintId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Complaint tracking ID (e.g., KSC0001)",
            "example": "KSC0001"
          }
        ],
        "responses": {
          "200": {
            "description": "Complaint details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "complaint": {
                          "$ref": "#/components/schemas/Complaint"
                        },
                        "statusHistory": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StatusLog"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Complaint not found"
          }
        }
      }
    },
...
    "/api/guest/service-request": {
      "post": {
        "summary": "Submit a guest service request",
        "tags": [
          "Guest"
        ],
        "description": "Submit a service request as a guest user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service request submitted successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/guest/verify-service-otp": {
      "post": {
        "summary": "Verify OTP for service request",
        "tags": [
          "Guest"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully"
          },
          "400": {
            "description": "Invalid or expired OTP"
          }
        }
      }
    },
    "/api/guest/track-service/{requestId}": {
      "get": {
        "summary": "Track service request status",
        "tags": [
          "Guest"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "requestId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Service request ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Service request details retrieved"
          },
          "404": {
            "description": "Service request not found"
          }
        }
      }
    },
...
    "/api/guest/files/{filename}": {
      "get": {
        "summary": "Serve uploaded guest files",
        "tags": [
          "Guest"
        ],
        "description": "Retrieve uploaded files from guest complaints",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the uploaded file"
          }
        ],
        "responses": {
          "200": {
            "description": "File served successfully",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
...
  }
}